Xbox Kodi 19.3 WORKAROUND
EDIT: USE AT YOUR OWN RISK. This will be have to be deleted AFTER kodi gets updated again so it can use the official version. There's instructions on how to delete it but I can't figure it out. You may have to uninstall and reinstall kodi.

EDIT 2: They said having the workaround when the new update comes out won't break anything so if you absolutely can't wait for the update then try this

Found this on the kodi website and have tried it myself. Currently watching something on Seren